requirements.txt
psycopg2==2.9.11
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
python-dotenv==1.2.1
redis==7.4.0
pymongo==4.16.0
dnspython==2.8.0
alembic==1.18.4
uvicorn==0.44.0
fastapi==0.135.3